Split Video.cpp to DirectVideo.cpp to remove most of the Win32 specific functions.
