+++ source/Disk.cpp
@@ -43,6 +43,7 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 #include "Registry.h"
 #include "SaveState.h"
 #include "Video.h"
+#include "DirectVideo.h"
 #include "YamlHelper.h"
 
 #include "../resource/resource.h"
