+++ source/Speaker.cpp
@@ -35,7 +35,7 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 #include "Log.h"
 #include "Memory.h"
 #include "SoundCore.h"
-#include "Video.h"	// VideoRedrawScreen()
+#include "DirectVideo.h"	// VideoRedrawScreen()
 #include "YamlHelper.h"
 #include "Riff.h"
